[Rework] Start rewriting of the redis cache logic
